Refactor temporal client
