+++ server/temporal-client/index.js
@@ -1,0 +1,2 @@
+const { TemporalClient } = require('./temporal-client');
+module.exports = { TemporalClient };
